updated ToricPointGenerator

+++ cymetric/pointgen/nphelper.py
@@ -36,10 +36,10 @@ def conf_to_monomials(conf):
         >>> monomials = conf_to_monomials(conf)
 
     Args:
-        conf (ndarray([nProj,nHyper], np.int)): Configuration matrix.
+        conf (ndarray([nProj,nHyper], int)): Configuration matrix.
 
     Returns:
-        list(nHyper,ndarray([nMonomials, nVars], np.int)): Monomial basis for
+        list(nHyper,ndarray([nMonomials, nVars], int)): Monomial basis for
             each hypersurface.
     """
     ambient = np.sum(conf, axis=-1)-1
@@ -206,17 +206,17 @@ def get_all_patch_degrees(glsm_charges, patch_masks):
     that the largest coordinates will be 1+0.j.
 
     Args:
-        glsm_charges (ndarray([nscaling, ncoords], np.int)): GLSM charges.
+        glsm_charges (ndarray([nscaling, ncoords], int)): GLSM charges.
         patch_masks (ndarray([npatches, ncoords], bool)): Patch masks with
             True at each coordinates, which is not allowed to vanish.
 
     Returns:
-        ndarray([npatches, ncoords, ncoords], np.int): degrees
+        ndarray([npatches, ncoords, ncoords], int): degrees
     """
     npatches, ncoords = np.shape(patch_masks)
-    all_patch_degrees = np.zeros((npatches, ncoords, ncoords), dtype=np.int)
+    all_patch_degrees = np.zeros((npatches, ncoords, ncoords), dtype=int)
     for i in range(npatches):
-        all_patch_degrees[i] = np.eye(ncoords, dtype=np.int)
+        all_patch_degrees[i] = np.eye(ncoords, dtype=int)
         patch_coords = np.where(patch_masks[i])[0]
         for j in range(ncoords):
             factors = np.linalg.solve(
@@ -224,7 +224,7 @@ def get_all_patch_degrees(glsm_charges, patch_masks):
             if not np.allclose(factors, np.round(factors)):
                 print('WARNING GLSM: NO INTEGER COEFFICIENTS.')
             for l, k in enumerate(patch_coords):
-                all_patch_degrees[i, j, k] -= np.round(factors[l]).astype(np.int)
+                all_patch_degrees[i, j, k] -= np.round(factors[l]).astype(int)
     return all_patch_degrees
 
 
@@ -234,7 +234,7 @@ def compute_all_w_of_x(patch_degrees, patch_masks, dim_cy = 3):
     homogeneous ambient space coordinates.
 
     Args:
-        patch_degrees (ndarray([npatches, ncoords, ncoords], np.int)): See also
+        patch_degrees (ndarray([npatches, ncoords, ncoords], int)): See also
             :py:func:`get_all_patch_degrees()`.
         patch_masks (ndarray([npatches, ncoords], bool)): Patch masks with
             True at each coordinates, which is not allowed to vanish.
@@ -245,11 +245,11 @@ def compute_all_w_of_x(patch_degrees, patch_masks, dim_cy = 3):
     """
     npatches, ncoords = np.shape(patch_masks)    
     w_of_x = np.zeros(
-        (ncoords, npatches, npatches, dim_cy, dim_cy), dtype=np.int)
+        (ncoords, npatches, npatches, dim_cy, dim_cy), dtype=int)
     del_w_of_x = np.zeros(
-        (ncoords, npatches, npatches, dim_cy, dim_cy, dim_cy), dtype=np.int)
+        (ncoords, npatches, npatches, dim_cy, dim_cy, dim_cy), dtype=int)
     del_w_of_z = np.zeros(
-        (ncoords, npatches, npatches, dim_cy, dim_cy, ncoords), dtype=np.int)
+        (ncoords, npatches, npatches, dim_cy, dim_cy, ncoords), dtype=int)
     # TODO: Add a warning for when the array becomes too large.
     # NOTE: There will be many zeros.
     for i in range(ncoords):
@@ -270,9 +270,9 @@ def compute_all_w_of_x(patch_degrees, patch_masks, dim_cy = 3):
                         patch_degrees[j][g1mask].T, v, rcond=None)
                     if not np.allclose(coeff, np.round(coeff)):
                         print('WARNING W(X): NO INTEGER COEFFICIENTS.')
-                    w_of_x[i, j, k, l] = np.round(coeff).astype(np.int)
+                    w_of_x[i, j, k, l] = np.round(coeff).astype(int)
                     # compute the derivative wrt to the g1 coordinates
-                    del_w_of_x[i, j, k, l] = w_of_x[i, j, k, l] - np.eye(dim_cy, dtype=np.int)
+                    del_w_of_x[i, j, k, l] = w_of_x[i, j, k, l] - np.eye(dim_cy, dtype=int)
                     # re-express everything in terms of degrees of the homogeneous
                     # ambient space coordinates
                     for m in range(dim_cy):
